requirements.txt
alembic==1.16.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
async-timeout==5.0.1
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==45.0.4
Django==5.1.3
django-allauth==65.9.0
django-ckeditor==6.7.3
django-filter==24.3
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-resized==1.0.3
django_celery_results==2.6.0
djangorestframework==3.15.2
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.115.7
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jwcrypto==1.5.6
kombu==5.5.4
Mako==1.3.10
Markdown==3.7
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pillow==11.0.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-jose==3.5.0
python-jwt==4.1.0
redis==4.5.5
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.2
starlette==0.45.2
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
websockets==14.2
